[HQ-535] Only remove alerts if necessary

+++ src/org/hyperic/hq/events/server/session/AlertDefinitionManagerEJBImpl.java
@@ -187,7 +187,9 @@ public class AlertDefinitionManagerEJBImpl
         
         // Delete the alerts
         watch.markTimeBegin("deleteByAlertDefinition");
-        dao.deleteByAlertDefinition(alertdef);
+        if (force) {
+            dao.deleteByAlertDefinition(alertdef);
+        }
         watch.markTimeEnd("deleteByAlertDefinition");
 
         // Remove the conditions
